+++ README.md
@@ -1,1 +1,6 @@
 -- MyBatis
+# setting datasource and config mybatis in application.yml file within resources directory
+# create mybatis-configuration.xml file within resources directory
+# create mapper.sql folder within resources directory (place which create xml file to communicate with DB)
+# create model package ~ table in DB (same level with MyappApplication.java)
+# create mapper package ~ map functions to file xml
