+++ src/main/java/com/thanhnd101/enrich/springboot/myapp/rest/UserController.java
@@ -1,7 +1,7 @@
 package com.thanhnd101.enrich.springboot.myapp.rest;
 
-import com.thanhnd101.enrich.springboot.mapper.UserMapper;
-import com.thanhnd101.enrich.springboot.model.User;
+import com.thanhnd101.enrich.springboot.myapp.mapper.UserMapper;
+import com.thanhnd101.enrich.springboot.myapp.model.User;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,24 +10,21 @@ import org.springframework.web.bind.annotation.RestController;
 
 /**
  * user mapper.
- *
- * @author thanhnd101
  */
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api")
-public class FunRestController {
+public class UserController {
 
   private final UserMapper userMapper;
 
   /**
    * get all users.
+   *
    * @return list users.
    */
   @GetMapping("/users")
-  public int getUsers() {
-    List<User> listUser = userMapper.findAll();
-
-    return listUser.size();
+  public List<User> getUsers() {
+    return userMapper.findAll();
   }
 }
