+++ src/main/resources/com.thanhnd101.mapper.sql/userMapper.xml
@@ -1,12 +1,12 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.thanhnd101.enrich.springboot.mapper.UserMapper">
-  <select id="findAll" resultType="com.thanhnd101.enrich.springboot.model.User">
-    SELECT id FROM USER
+<mapper namespace="com.thanhnd101.enrich.springboot.myapp.mapper.UserMapper">
+  <select id="findAll" resultType="com.thanhnd101.enrich.springboot.myapp.model.User">
+    SELECT * FROM `es-java`.`user`
   </select>
   <insert id="createUser">
-    INSERT INTO USER (id, username, password, email, address, birthday)
-    VALUES ('SGSDSSARASF', 'Tom B. Erichsen', 'atetwerwfer', 'adasda@gmail.com', '4006', '2020/12/21');
+    insert into `es-java`.`user` (id, username, password, email, address, birthday)
+    values('ABCD', 'Tom B. Erichsen', 'atetwerwfer', 'adasda@gmail.com', '4006', '2020-12-21');
   </insert>
 </mapper>
